+++ sync-for-commerce/RELEASES.md
@@ -571,3 +571,13 @@ Based on:
 - [typescript v2.1.0] sync-for-commerce
 ### Releases
 - [NPM v2.1.0] https://www.npmjs.com/package/@codat/sync-for-commerce/v/2.1.0 - sync-for-commerce
+
+## 2023-09-21 00:07:22
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.88.1 (2.122.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.1] sync-for-commerce
+### Releases
+- [NPM v2.1.1] https://www.npmjs.com/package/@codat/sync-for-commerce/v/2.1.1 - sync-for-commerce